fix: compile error in writeSnapshot
This is probably due to the component upgrades (maybe the old versions missed this or just ignored it). Anyways, as far as I understand this, there is very little point to call writeBytes with null data, and the kotlin type safety wont allow it anyways. If it would, this would probably be npe.