fix: remove compile error in fetchPlayerInfo
This is much trickier: not quite sure why we juggle the player data like this. Pretty much all the data is available in the player object anyways, so thought I would just return that for now. I know it breaks the API, but my purpose for now is to get this show on the road and we are not using this feature ourselves (yet). I will probably come back to this either by 1) forced when something breaks or 2) when we actually have time. :D